Found solution for disparate entry points for data using From & Into
When constructing a memory region from an integer, there expressly needs
to be a way to treat that integer literal as a memory region so that it
will be initialized as written (0b101 means 10100000 not 00000101).

There will be a Mem idiomatic type and a Num subclass to handle From and
Into appropriately. This can be extended to I32, F32, etc.