Removing `Num` type (reasons below)
The `Num` type worked off of the premise that it could store both
positive and negative big integers while still using two's-complement
encoding.

It turned out that this is not useful because `Num[1](1)` was truly both
a semantic error and a bad default feature because unsigned 1 bit
numbers need to be able to store `0` or `1` but 1 bit signed numbers can
only denote `+0` and `-0`.

This mish-mash of semantics is not sensible for a default number type so
`Num` will be split out into `Unsigned` and `Signed` so that 1 bit
numbers can be supported properly.